% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_package_data}
\alias{generate_package_data}
\title{Generate and Save Package Data}
\usage{
generate_package_data()
}
\description{
This function generates and saves all the datasets needed for the tutorial.
It should be run during package development, not by end users.
}
