% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_multilevel_data}
\alias{get_multilevel_data}
\title{Access the multilevel penguins example dataset}
\usage{
get_multilevel_data()
}
\value{
A dataset suitable for multilevel modeling
}
\description{
Access the multilevel penguins example dataset
}
