% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_nonlinear_data}
\alias{get_nonlinear_data}
\title{Access the nonlinear example dataset}
\usage{
get_nonlinear_data()
}
\value{
A dataset with a nonlinear relationship
}
\description{
Access the nonlinear example dataset
}
