% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_penguins_clean}
\alias{get_penguins_clean}
\title{Access the clean penguins dataset}
\usage{
get_penguins_clean()
}
\value{
The penguins dataset with NA values removed
}
\description{
Access the clean penguins dataset
}
