% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_penguins_sex}
\alias{get_penguins_sex}
\title{Access the penguin sex prediction dataset}
\usage{
get_penguins_sex()
}
\value{
A dataset for binary classification
}
\description{
Access the penguin sex prediction dataset
}
