% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_to_brms.R
\docType{data}
\name{intro_to_brms}
\alias{intro_to_brms}
\title{Introduction to Bayesian Regression with brms}
\description{
This tutorial provides an interactive introduction to Bayesian regression modeling
using the brms package. It covers the fundamentals of Bayesian statistics,
model specification, fitting, interpretation, and comparison using brms.
}
\examples{
if (interactive()) {
  learnr::run_tutorial("intro_to_brms", package = "brmsTutorial")
}
}
\keyword{tutorials}
