% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_to_brms.R
\docType{data}
\name{nonlinear_data}
\alias{nonlinear_data}
\title{Nonlinear example data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{x}{predictor variable}
\item{y}{response variable, following an exponential decay with noise}
}
}
\usage{
nonlinear_data
}
\description{
A dataset with a nonlinear relationship between x and y.
}
\keyword{datasets}
