% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_to_brms.R
\docType{data}
\name{penguins_clean}
\alias{penguins_clean}
\title{Palmer Penguins Dataset (cleaned)}
\format{
A data frame with penguin measurements:
\describe{
\item{species}{a factor denoting penguin species (Adelie, Chinstrap, Gentoo)}
\item{island}{a factor denoting island in Palmer Archipelago, Antarctica}
\item{bill_length_mm}{a number denoting bill length (millimeters)}
\item{bill_depth_mm}{a number denoting bill depth (millimeters)}
\item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
\item{body_mass_g}{an integer denoting body mass (grams)}
\item{sex}{a factor denoting penguin sex (female, male)}
\item{year}{an integer denoting study year (2007, 2008, 2009)}
}
}
\source{
Horst AM, Hill AP, Gorman KB (2020). palmerpenguins: Palmer
Archipelago (Antarctica) penguin data. R package version 0.1.0.
https://allisonhorst.github.io/palmerpenguins
}
\usage{
penguins_clean
}
\description{
A cleaned version of the Palmer Penguins dataset with NA values removed.
}
\keyword{datasets}
