% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_to_brms.R
\docType{data}
\name{penguins_sex}
\alias{penguins_sex}
\title{Penguin sex prediction dataset}
\format{
A data frame with penguin measurements:
\describe{
\item{sex}{binary factor (female/male)}
\item{bill_depth_mm}{bill depth in millimeters}
\item{bill_length_mm}{bill length in millimeters}
}
}
\source{
Horst AM, Hill AP, Gorman KB (2020). palmerpenguins: Palmer
Archipelago (Antarctica) penguin data. R package version 0.1.0.
}
\usage{
penguins_sex
}
\description{
A subset of the Palmer Penguins dataset for demonstrating logistic regression with binary outcomes.
}
\keyword{datasets}
